requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-multipart==0.0.9
google-generativeai==0.7.2
pymupdf==1.24.2
python-docx==1.1.0
